+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/Drawiner-master.iml" filepath="$PROJECT_DIR$/.idea/Drawiner-master.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/Drawiner.iml" filepath="$PROJECT_DIR$/.idea/Drawiner.iml" />
     </modules>
   </component>
 </project>
