+++ Canvas.py
@@ -15,14 +15,16 @@ class Game(Widget):
     def build(self):
         self.usership.bind(on_coords=self.move)
 
-    def get_main_sky(self, found, *args):
+    @staticmethod
+    def get_main_sky(found, *args):
         for img in args:
             if ((img.pos[0] <= found.center_x <= img.pos[0] + img.size[0]) and
                     (img.pos[1] <= found.center_y <= img.pos[1] + img.size[1])):
                 return img
         return -1
 
-    def get_quarter(self, source, found):
+    @staticmethod
+    def get_quarter(source, found):
         if ((source.center_x <= found.center_x) and
                 (source.center_y <= found.center_y)):
             return 1
@@ -41,11 +43,11 @@ class Game(Widget):
     def move(self):
         if len(self.skyboxes) == 0:
             self.skyboxes = [self.skybox_1, self.skybox_2, self.skybox_3, self.skybox_4]
-        main_sky = self.get_main_sky(self.usership, self.skybox_1,
+        main_sky = Game.get_main_sky(self.usership, self.skybox_1,
                                      self.skybox_2, self.skybox_3,
                                      self.skybox_4)
         self.skyboxes.remove(main_sky)
-        quarter = self.get_quarter(main_sky, self.usership)
+        quarter = Game.get_quarter(main_sky, self.usership)
         if quarter != self.last_quarter:
             if quarter == 1:
                 self.skyboxes[0].coords.x = main_sky.coords.x
