+++ UserShip.py
@@ -1,5 +1,6 @@
 from Ship import Ship
 from AbstractObjects import Singleton
+from kivy.vector import Vector
 
 
 class UserShip(Ship, Singleton):
@@ -17,5 +18,32 @@ class UserShip(Ship, Singleton):
         "right_t": [0, 0, 0],
         "weapon_1": [0, 0, 0]
     }
-    # inventory = Inventory()
+
     source = 'img/usership.png'
+
+    def thrust(self, direction):
+
+        if self.velocity.length() > self.max_speed:
+            self.velocity = self.velocity.normalize()*self.max_speed
+            return
+
+        if direction == "forward_t":
+            self.velocity = (self.velocity
+                             + Vector(0.1, 0).rotate(self.angle)
+                             * self.heat[direction][0])
+            self.overheat(direction)
+        elif direction == "backward_t":
+            self.velocity = (self.velocity
+                             + Vector(0.05, 0).rotate(self.angle + 180)
+                             * self.heat[direction][0])
+            self.overheat(direction)
+        elif direction == "left_t":
+            self.velocity = (self.velocity
+                             + Vector(0.03, 0).rotate(self.angle + 90)
+                             * self.heat[direction][0])
+            self.overheat(direction)
+        elif direction == "right_t":
+            self.velocity = (self.velocity
+                             + Vector(0.03, 0).rotate(self.angle + 270)
+                             * self.heat[direction][0])
+            self.overheat(direction)
